Collection UI Started
Started basic inventory UI stuff for the collection page
Made changes to the other scripts to improve counting how many items and adding the UI Container of the item.
It seems item.itemCount does carry through, hopefully, need to further test with remove items and crafting system.
Crafting system may need another JSON just for recipes.
What else needs to be databased?
Also made the old panel inventory test just reference now, and may need to remove later on when collection page has been proven to work without much errors - to which a new event in UIEventHandler proven the old event to be useless

+++ Gacha Shop/Assets/Scripts/PlayerSystem.cs
@@ -6,7 +6,7 @@ using TMPro;
 
 public class PlayerSystem : MonoBehaviour
 {
-    public float money;
+    public double money;
     // public List<GameObject> items;
     public TextMeshProUGUI timesX;
     public TextMeshProUGUI moneyCounter;
